split CoreValueCard from CoreValuesBlock component
